GitHub HR Hack
Conditionally insert an horizontal rule before the example block at the
end of the document when previewed on GitHub. This is required to add
some visual separation before editor notes because example blocks are
not styled on GitHub.

Improve handling of PMNG version in documents Header block via revremark
attribute.